Factor out CommentBox.js
